Fix compatibility with Docker 1.9 - custom network and no container links
Docker 1.9 changes default networking behaviour, to reference each other by
container name they need to be on a non-default network. Added variable in
container_functions.sh along with some functions to create a new network.

Added all of the containers to this network.

It's not possible to use --links with custom networks so have added a few
environment variables passed in at run time to provide the same data previously
linked in. Behaviour of gerrit-entrypoint no longer suitable, so not using
built in postgresql setup, instead setting this ourselves in config.sh now