Add a containerized redis for build-debian-contextualizer & apt-version-mapper
Added a new container containing redis for use by build-debian-contextualizer
and apt-version-mapper. It does not require volume mounting as the data it
stores is only used temporarily and it doesn't require any changes to the
default config. As such redis for us doesn't have a build step and instead
just uses the official redis image as is.