Initial work on Bugzilla container
There wasn't a suitable pre-existing Dockerfile so we're aiming to make
our own. This initial commit contains a fair chunk but there is still a
lot of configuration needed to automate Apache2 being happy & setup
correctly.

Need to automate:
Pg-gerrit creating the bugs user - Running into issues of not being able
to pre-supply a password using createuser

Configuration is correct - docker-entrypoint.sh in Bugzilla is a bit
tempremental, I think this problem mostly goes away once we integrate it
with mkbuildsystem and co.

Set up Admin email/Name - Currently we are supplying ADMIN_EMAIL, ADMIN_REALNAME
and ADMIN_PASSWORD but the configuration is still asking us to manually
enter these as if they are missing. Suspect there is a syntax issue.

Start apache and configure its config - Auto-configure and auto-start apache2.
We have started some of this by moving bugzilla to /var/www/. Need to change
DocumentRoot and enable mod's for cgi's to work (+more, just touched the
surface)

After this the image should be mostly ready, then we need to also:
Integrate the whole thing with mkbuildsystem.
Fix in hooks and such
Possibly submit this upstream(???)