Bugzilla -> Gerrit hooks mostly configured
It will automatically come up with working Hooks, just need to commit and
+2 and you will see it do work.

It is being odd and firing/detecting automatic merge's on EVERY commit so far
which we are trying to track the why's and how to fix.

Reshuffled some of the container_functions patch* func's.

Changed Bugzilla to the same version as live (and small tweaks to get it
happy and working again subsequently)

Added a patch for change-merged to support the new version of Gerrit we are
using.

Various other small tweaks, including move sudo.txt to Gerrit instead of
Jenkins. All small fixe's towards getting Gerrit hooks working.