Jenkins: Fix Buildlog's to point to the new Buildfs container
Added some extra config to pass Jenkin's config.xml the URL for Buildfs
container. Currently this uses localhost but it is all parametised so changing
if/when host configuration changes is easy to do.

Some small name changes to more accurately reflect what they are/what they
are used for.

Fixed a typo in an earlier commit (build_fs instead of buildfs)