Fix dev imports, so you can bring up a local copy w/ a subset of data
Replacing the old mechanism of files_to_copy there is a subset of data
which exists in /global/users/jonathan.barron/cbuildsystem-starter-pack.

Import.sh with the dev option now copies /container-mounts and /db_dumps
from there into the local $BUILD_HOME. After that the process for importing
into db and reloading is the same as bringing up a live one.
the same