Remove unused / redundant functions from Container_functions.sh
We used to have a lot of functions which would copy a local "files_to_copy"
into the mounted filesystems. Now that this is handled by import.sh they
are not needed.

In addition the functions we were using to patch a local copy of buildsystem
are no longer needed as we are using a topic branch with the changes already
in.