Add Postfix container to handle Gerrit's email notifications
Using an off the shelf image (catatnight/docker-postfix) and a little bit of
config tweaking the gerrit-container now sends emails similar to live gerrit.

I have had to set the gerrit user.email (in its config) to a hard coded value,
as otherwise it defaults to <user>@<hostname> and the hostname for containers
is a generated string.