Static SSH /etc/ssh/ hostkeys files baked into container images.
Previously a new host key was being generated whenever a new container
was built. A set of four key collections has been generated and saved
in the cbuild-secrets repository and is now copied into the Gerrit
and Jenkins images whenever thay are rebuilt.

Individual keys are stored and used for the two jenchild images.

Therefore, we now create two distinct images rather than just one
image instantiated twice as we did before. This means that if we want
to scale up a new image would need to be generated.

Change-Id: I7d2d861c9c66b5a9243ad920e6d02efd4c9f4955