Redo: Volume mount /var/gerrit/review_site to host so ssh_host_key persists
We previously did this, but removed it as the data inside this directory was
deemed working data and held no value.

There is a long standing issue with host keys being re-generated on start/stop.
Other commits have fixed /etc/ssh and /var/gerrit/.ssh to persist but gerrit
generates on gerrit init a special ssh_host_key stored in review_site/etc.

So we've remounted the review_site directory. A possible advantage to this,
we may be able to rework reindexing behaviour now to reduce start-up times.