Create initial README.md's for each Docker image we use
I have written up some notes / tips / examples for each of the docker
images that we use. Redis & Postfix are in the gerrit one, since they
do not have a /directory or Dockerfile as they're used as is.

There is likely things missed, or incorrect, or that will  need updating
as times change but for now this will hopefully be useful for people to
get familiar with how this all works.