Changes ready for files now being on Buildfs
Defaults has been changed to reflect the new folder structure:

On buildfs we've made /srv/container_data/* which have been nfs mounted
on buildwhale1 to /usr/src/buildfs-mounts/*

We have a script prepared and ready to copy the contents of /usr/src/
container-mounts onto buildfs.

Existing on buildfs already were /srv/build (/mnt/build) and /home/build.

We use the same /srv/build as metal buildsystem but /home/build we have
a container replica because ssh config's and stuff need to be different